+++ src/Undefinable/orElse.ts
@@ -1,7 +1,7 @@
-import { Undefinable, isNotUndefined } from './Undefinable';
+import { Undefinable } from './Undefinable';
 
 export function orElseForUndefinable<T>(a: Undefinable<T>, b: (this: void) => Undefinable<T>): Undefinable<T> {
-    if (isNotUndefined(a)) {
+    if (a !== undefined) {
         return a;
     }
     else {