+++ src/Undefinable/unwrapOrElse.ts
@@ -1,7 +1,7 @@
-import { Undefinable, isNotUndefined } from './Undefinable';
+import { Undefinable } from './Undefinable';
 import { RecoveryFn } from '../utils/Function';
 
 export function unwrapOrElseFromUndefinable<T>(v: Undefinable<T>, def:  RecoveryFn<T>): T {
-    return isNotUndefined(v) ? v : def();
+    return (v !== undefined) ? v : def();
 }