+++ src/PlainOption/andThen.ts
@@ -1,11 +1,11 @@
-import { Option, isSome } from './Option';
+import { Option } from './Option';
 
 /**
  *  XXX:
  *  Some languages call this operation flatmap.
  */
 export function andThenForOption<T, U>(src: Option<T>, fn: (this: void, v: T) => Option<U>): Option<U> {
-    if (isSome(src)) {
+    if (src.ok) {
         const r = fn(src.val);
         return r;
     }