+++ src/PlainOption/mapOrElse.ts
@@ -1,9 +1,9 @@
 import { MapFn, RecoveryFn } from '../utils/Function';
-import { Option, isSome } from './Option';
+import { Option } from './Option';
 
 export function mapOrElseForOption<T, U>(src: Option<T>, def: RecoveryFn<U>, selector: MapFn<T, U>): U {
     let r: U;
-    if (isSome(src)) {
+    if (src.ok) {
         r = selector(src.val);
     }
     else {