+++ src/Maybe/and.ts
@@ -1,5 +1,5 @@
-import { isNotNullAndUndefined, Maybe } from './Maybe';
+import { Maybe } from './Maybe';
 
 export function andForMaybe<T, U>(a: Maybe<T>, b: Maybe<U>): Maybe<U> {
-    return isNotNullAndUndefined(a) ? b : a;
+    return (a !== undefined && a !== null) ? b : a;
 }