+++ src/Maybe/do.ts
@@ -1,8 +1,8 @@
 import { DoFn } from '../utils/Function';
-import { Maybe, isNullOrUndefined } from './Maybe';
+import { Maybe } from './Maybe';
 
 export function doOnMaybe<T>(v: Maybe<T>, fn: DoFn<T>): void {
-    if (isNullOrUndefined(v)) {
+    if (v === undefined || v === null) {
         return;
     }