+++ src/Maybe/map.ts
@@ -1,10 +1,10 @@
 import { expectNotNullAndUndefined } from './expect';
 import { ERR_MSG_SELECTOR } from './ErrorMessage';
 import { MapFn } from '../utils/Function';
-import { Maybe, isNotNullAndUndefined } from './Maybe';
+import { Maybe } from './Maybe';
 
 export function mapForMaybe<T, U>(src: Maybe<T>, selector: MapFn<T, U>): Maybe<U> {
-    if (isNotNullAndUndefined(src)) {
+    if (src !== undefined && src !== null) {
         const r = selector(src);
         return expectNotNullAndUndefined(r, ERR_MSG_SELECTOR);
     }