+++ src/Maybe/mapOrElse.ts
@@ -1,12 +1,12 @@
 import { expectNotNullAndUndefined } from './expect';
 import { ERR_MSG_MUST_NOT_RETURN_NULL_OR_UNDEF, ERR_MSG_SELECTOR } from './ErrorMessage';
 import { MapFn, RecoveryFn } from '../utils/Function';
-import { Maybe, isNotNullAndUndefined } from './Maybe';
+import { Maybe } from './Maybe';
 
 export function mapOrElseForMaybe<T, U>(src: Maybe<T>, def: RecoveryFn<U>, selector: MapFn<T, U>): U {
     let r: U;
     let msg = '';
-    if (isNotNullAndUndefined(src)) {
+    if (src !== undefined && src !== null) {
         r = selector(src);
         msg = ERR_MSG_SELECTOR;
     }