+++ src/Maybe/orElse.ts
@@ -1,7 +1,7 @@
-import { Maybe, isNotNullAndUndefined } from './Maybe';
+import { Maybe } from './Maybe';
 
 export function orElseForMaybe<T>(a: Maybe<T>, b: (this: void) => Maybe<T>): Maybe<T> {
-    if (isNotNullAndUndefined(a)) {
+    if (a !== undefined && a !== null) {
         return a;
     }
     else {