+++ src/Maybe/unwrapOr.ts
@@ -1,5 +1,5 @@
-import { Maybe, isNotNullAndUndefined } from './Maybe';
+import { Maybe } from './Maybe';
 
 export function unwrapOrFromMaybe<T>(v: Maybe<T>, def: T): T {
-    return isNotNullAndUndefined(v) ? v : def;
+    return (v !== undefined && v !== null) ? v : def;
 }